# chore: upgrading vite to newer version