--- a/frontend/cypress.config.ts
+++ b/frontend/cypress.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
             on(
                 'file:preprocessor',
                 vitePreprocessor({
-                    configFile: path.resolve(__dirname, './vite.config.ts'),
+                    configFile: path.resolve(__dirname, './vite.config.mts'),
                     mode: 'development',
                 }),
             );