--- a/frontend/src/component/feature/FeatureStrategy/FeatureStrategyCreate/FeatureStrategyCreate.test.tsx
+++ b/frontend/src/component/feature/FeatureStrategy/FeatureStrategyCreate/FeatureStrategyCreate.test.tsx
@@ -10,11 +10,11 @@ test('formatAddStrategyApiCode', () => {
             'unleashUrl',
         ),
     ).toMatchInlineSnapshot(`
-      "curl --location --request POST 'unleashUrl/api/admin/projects/projectId/features/featureId/environments/environmentId/strategies' \\\\
-          --header 'Authorization: INSERT_API_KEY' \\\\
-          --header 'Content-Type: application/json' \\\\
+      "curl --location --request POST 'unleashUrl/api/admin/projects/projectId/features/featureId/environments/environmentId/strategies' \\
+          --header 'Authorization: INSERT_API_KEY' \\
+          --header 'Content-Type: application/json' \\
           --data-raw '{
-        \\"id\\": \\"strategyId\\"
+        "id": "strategyId"
       }'"
     `);
 });