--- a/frontend/tsconfig.node.json
+++ b/frontend/tsconfig.node.json
@@ -4,5 +4,5 @@
     "module": "esnext",
     "moduleResolution": "node"
   },
-  "include": ["vite.config.ts"]
+  "include": ["vite.config.mts"]
 }