<!-- Thanks for creating a PR! To make it easier for reviewers and everyone else to understand what your changes relate to, please add some relevant content to the headings below. Feel free to ignore or delete sections that you don't think are relevant. Thank you! ❤️ -->

## About the changes

Upgrading Vite to 5 to speed up build time (used by enterprise install) and test time (for PRs in this repo).

Other changes:
* snapshot format changed
* vite config should have mts extension (https://vitejs.dev/guide/migration.html#deprecate-cjs-node-api)

New test: 76 seconds

Old test: 88 seconds

New build: 36 seconds

Old build: 46 seconds

### Important files
<!-- PRs can contain a lot of changes, but not all changes are equally important. Where should a reviewer start looking to get an overview of the changes? Are any files particularly important?  -->


## Discussion points
<!-- Anything about the PR you'd like to discuss before it gets merged? Got any questions or doubts? -->